SocialProfile: bug #27032 - new EditProfile special page to edit other users' social profiles. Patch by Frozen Wind <tuxlover684@gmail.com> with some changes by me.
In an attempt to reduce the amount of necessary code duplication, I made SpecialEditProfile subclass SpecialUpdateProfile. Sadly this also means that in SpecialUpdateProfile.php, on lines 141, 144 and 147 we have to pass $wgUser to the functions in order not to throw out E_STRICT warnings when the user is editing their own social profile via Special:UpdateProfile. Special:EditProfile is now also listed on Special:SpecialPages instead of being an UnlistedSpecialPage.