SocialProfile: allow specifying and showing the user's year of birth on social profile pages. Patch by Rinat Silnov.
To use this feature, go to Special:UpdateProfile/preferences, toggle the new "Show year of birth" option, save preferences, then add your birthday on Special:UpdateProfile, save again and now your birthday will be shown on your profile.

Do note that because of the mw.loader.using() call in UserProfile/UpdateProfile.js, this breaks backwards compatibility with pre-ResourceLoader MediaWikis (that is, 1.16 and older). I don't know if this works with 1.17, but I tested this against the current 1.18 branch.