Add some common/shared assets to be used by PictureGame, PollNY & QuizGame
LightBox.js is based on the version used by QuizGame, although the on DOM
ready call to the init() method has been removed and the code has been
JSHinted.

flash.js is a new thing to be used by LightBox consumers so that they can
display either ajax-loading.swf for Flash-enabled clients or
ajax-loader.gif or ajax-loader-white.gif for clients which do not support
Flash, such as most mobile devices.

Bug: T100025
Change-Id: I287b4967ffb78dbd4f30a25ae7f2e029c23ee312