Allow everyone to remove their own avatar via Special:RemoveAvatar
Previously users without the 'avatarremove' user right couldn't remove
their own avatar cleanly and their only option (besides asking a
privileged user to remove the avatar) was to save a copy of the default
avatar and upload it to the wiki via Special:UploadAvatar. Obviously this
was far from ideal as the workflow was unnecessarily complex *and* it
resulted in lots of duplicated files on the server.

Bug: T158713
Change-Id: I4e16ab42b6101b7301d04a896d98e99bd91e306a