Add a class to <body> on NS_USER pages depending on whether the user has chosen the social profile or wikitext user page as their default User: page
If the user has chosen the wikitext user page, the "mw-wiki-user-page"
class is appended to the <body> element; for users who have chosen the
social profile page instead, the class is "mw-social-profile-page".
This class is only applied on root user pages, e.g. [[User:Foo]], not on
subpages (e.g. [[User:Foo/Sandbox]], etc.) and obviously only for existing
users and only when users are allowed to choose their user page type
($wgUserPageChoice = true;, which happens to be the default).

Bug: T167506
Change-Id: I6f37f093e68c1949d0169ba4b4cd2ee9a9813a81