UserBoard i18n: Fix phrasing of Echo notification bundle message
This message has been phrased incorrectly, as it wrongly assumes
that each message is sent by a different person. The easiest way to
"fix" this error is to rephrase this message.
 - New: "{{PLURAL:$1|One message|$1
messages|100=99+ messages}} have been left on your user board."
 - Original: "{{PLURAL:$1|One person|$1 people|100=99+ people}} left
   messages on your user board."

Steps to reproduce:
1. Install the Echo extension on your local test wiki
2. Sign into one of your accounts and send multiple messages (e.g 3)
to another account of yours
3. Sign into your other account and visit [[Special:Notifications]],
and receive this wrongly phrased message.

Change-Id: Idcdeacf73a1479b7dbb0146736e0f3e72b81041f