Create a SystemGiftIcon object to replace the static SystemGifts::getGiftImage()
This new SystemGiftIcon object replaces the static function
'SystemGifts::getGiftImage()'. This is designed very similarly to the wAvatar
class in UserProfile.

With this new object, we are:
 - Reducing code duplication by having this object build the HTML itself too
 - Using the Html class in core instead of manual HTML string construction
 - Consistently providing an alt attribute which also uses the
   built-in 'ga-gift' i18n message; in some instances this alt attribute
   was a hard-coded English string or empty/non-existent.

For semantics, we also use the 'icon' term instead, since in this context these
are supposed to be small image representations of a gift.

Change-Id: Id9b41142efda0dfe0f02a88c3e33d4a97ae296e7