Don't load `ext.socialprofile.userprofile.diff` module twice
We originally had subscribers to both of these hooks:
- DifferenceEngineShowDiff
- DifferenceEngineShowDiffPage

The difference is that DifferenceEngineShowDiff has access to the
DifferenceEngine object (and access to OutputPage from there),
while DifferenceEngineShowDiffPage /only/ had access to OutputPage.

We remove our hook subscriber to DifferenceEngineShowDiffPage as
to improve performance by not loading this module twice.

Change-Id: I2427fba773a46d760dea4b3565055b52165e0ad1