Sniff: Enable 2 Squiz.Classes.* sniffs
Enables the following 2 related sniffs:
 - Squiz.Classes.SelfMemberReference.NotUsed
 - Squiz.Classes.ValidClassName.NotCamelCaps

The `wAvatar` class is the only class that starts with a lowercase
letter purely because of technical debt left from ArmchairGM. This
will eventually be fixed in some future follow-up patch.

We still want to have this sniff rule enabled, so we can do this by
adding an inline `phpcs:ignore` comment in wAvatar.

Change-Id: Ifdbaf12b8bf93e6f9f4e5578ecb0a856b6857007