Fix duplicate page title on Special:UploadAvatar and incorrect page title on Special:UpdateProfile/custom
The duplicate page titles were a leftover from the ArmchairGM days, when there was a global variable to control whether or not to output the usual <h1> page title element. Some extensions would set $wgSupressPageTitle = true; (sic) to hide the default <h1> element and generate it themselves. That global variable was killed for good a while ago and the whole idea was rather silly to begin with.

Change-Id: Ia6d59bcf5ae1988923f192d0eb300ad17d329cf8