Set mFileProps in UploadAvatar#verifyUpload to keep AbuseFilter happy
Normally mFileProps is set in UploadBase's verifyFile() or verifyPartialFile().
Other UploadBase code seems to blindly assume that "mFileProps is always set (to an array) and is never null", and because AbuseFilter hooks into the UploadStashFile hook and typehints $props to an array, uploading avatars breaks when $wgEmailConfirmToEdit is set to true and AbuseFilter is installed.
This works around that issue without commenting out the code in SpecialUploadAvatar#showRecoverableUploadError, because it feels slightly cleaner to me than flat-out commenting out code.

Bug: T239447
Change-Id: I0fd74324a7d1cc3ab133f0f68ce8c86c9305d394