Remove unused and broken UserSystemGifts#doesUserOwnGift
It was broken because it was obviously bad copypasta from UserGifts.
The user_system_gift table does *not* have a field called "ug_user_id_to" (but user_gift does).

The comparison should've been if ( $user_id == $s->sg_user_id ) instead, but because this method is unused and doesn't seem overly useful, I'm removing it.

Both UserSystemGifts and SystemGifts classes have a method called "doesUserHaveGift", and functionally those two are almost identical:
the UserSystemGifts method returns bool true on success or bool false on failure, whereas the SystemGifts method returns bool false on failure
and the ID of the requested system gift on success (i.e. when the user has that particular award).
FIXME: this is code duplication, obviously.

Change-Id: I0b4727be2929f3f257e7ca1ed81e7fb2a3b7f7e6