No more raw HTML in UserWelcome's i18n messages!
There is the additional side effect of the user levels help page now showing up as a redlink if the page doesn't exist;
previously it'd always show up as a bluelink since the link was built as a HTML snippet instead of by MW.
I don't think that's a big issue; quite the opposite, in fact, as I believe it'll encourage admins to create that page when it doesn't exist.

Also added/updated message documentation (qqq) while at it.

Change-Id: I531b7a8f69c32dbb45d7b3e7604c3fbd80fe237e