Always render User: pages as blue links, never as red links
Pages like Special:TopUsers etc. which use LinkRenderer would be slightly
confusing if they'd show a mixture of red and blue links when in fact,
regardless of the URL params, with SocialProfile installed they behave the
same.

This was not much of an issue back when such special pages and the like
used to use hand-written, raw HTML instead of LinkRenderer (or the older Linker).
Now things are different.

Change-Id: I7c8e2c229b73f354ed1ff351db194078a1173b3b