Don't call a nonexistent global function, wfImageArchiveDir, which hasn't existed since the MW 1.10 days (or so)
A more "proper" replacement for the function would be to use FileRepo methods, but that won't work for us because award and user-to-user gift  images are intentionally stored outside the usual MW file architecture, hence no FileRepo here.
The same is true for SportsTeams' sport and team logo images as well.

Bug: T198154
Change-Id: Ifd6b155a0e3e8147aa2628f11dab6715e8921804