Implement Special:UpdateEditCounts as a maintenance script
It uses about the same logic as SpecialUpdateEditCounts.php. However,
I've made some adjustements to the edit count logic.

First, traverse the list of users and actors. Unless on a wiki farm with
shared users, which may get lots of users without edits on the local
wiki, this should get only the relevant users and avoid querying the
revision table for anonymous and imported users.

Then, query the revision table, filtering out bots. I've added a
condition to avoid users where the editcount is 0 (which means there
will be no edits to find), avoiding the expensive query to the
revision table. There may be lots of users with editcount 0 due to
users creating accounts and never editing, or spambots with blocked
edits due to edit filters.

Since the logic should be shared by both the special page and the
maintenance script, I've created a new class that contains the logic
shared between them.

Bug: T341098
Change-Id: I5c01dcd3f863221a58132a0e9b42642a8affd94c