Don't fatal when trying to view the User:/User_profile: page of a username that cannot exist
Previously viewing "User:Ãƒ‰ã‚¤ãƒ„ã‚¤ãƒƒãƒ’ã¯æœ€å¼·ãƒ€ã‚¹" and "User_profile:Ãƒ‰ã‚¤ãƒ„ã‚¤ãƒƒãƒ’ã¯æœ€å¼·ãƒ€ã‚¹" would result in a fatal.
This happened because UserProfile's constructor is not tolerant of garbage (invalid) input, it assumes that it's only passed in a valid User object.
That can't obviously happen for a user name that's not valid and thus we end up with a fatal.

This still feels like a dirty hack more than a proper fix, but it's...something. At least it fixes the immediate problem.

Change-Id: I7f8e6775167174a543ec32ffaf2d9a6b303fe1c0