Unbreak CI/phan
This isn't a complete fix but more like baby steps in the right
direction by e.g. reducing the amount of messages accepting raw HTML.

Furthermore, I'm fairly confident that the entries in UserProfilePage are
NOT double-escaped for real, hence why I opted to suppress those warnings.

The UserActivity class still needs (a lot) more work so that it can pass
phan clearly without having to suppress anything.
But I suppose that at the end of the day, this is still an improvement over
the status quo, so there's that.

Change-Id: I56eab3edf6e52a48d6fca0b8a2b94ff735efedda