Remove nonsensical code that was probably bad copypasta from SpecialAddRelationship.php
SpecialAddRelationship.php has a basically identical line of code.
This line, despite having been in SocialProfile from the very beginning (of its public release as "SocialProfile" circa February 2008), is *not* present in the original ArmchairGM version of this file.
And it makes sense, why would you be *sending* a request from a page that's merely meant to allow you to _view_ requests sent to you???

Also fixes this phan-reported issue:
17:07:13 UserRelationship/includes/specials/SpecialViewRelationshipRequests.php:85 PhanTypeMismatchArgument Argument 1 ($userTo) is $this->user_name_to of type string but \UserRelationship::addRelationshipRequest() takes \User defined at UserRelationship/includes/UserRelationshipClass.php:41

Change-Id: Icf996f7885c8be58616f00aad77b50a5d9b8df24