"Fix" (suppress) both remaining PhanTypeArraySuspiciousNullable issues
17:07:13 UserGifts/includes/specials/SpecialGiftManager.php:369 PhanTypeArraySuspiciousNullable Suspicious array access to $gift of nullable type ?array|?non-empty-array<mixed,mixed>=
17:07:13 UserProfile/includes/avatar/Avatar.php:119 PhanTypeArraySuspiciousNullable Suspicious array access to $backend->getFileBackend()->getFileStat(['src'=>(($backend->getContainerStoragePath() . '/') . $avatar_filename)]) of nullable type array|false|null

I'm not sure about the avatar one but the GiftManager one is largely not relevant.
Granted, Gifts::getGift() *can* return an _empty_ array if it's passed in a non-numeric $id or there is no such gift with the given ID, but plenty of code, including right here in SpecialGiftManager.php on L345 onwards assumes that will never happen.
Should they? Probably not, no. Is it much of a concern? Also probably not. In any case, let's enable this phan check on a repo-wide scale and suppress just the current known two instances to ensure that no *new* errors of this type are introduced.

Change-Id: I74bafe213319adbca1a1a8d28a959ee49a15ed2f