Let $id params in SocialProfileFileBackend always be ints and cast them to string in getFileName()
Right now there are a total of ~40 phan issues.
The vast majority of these, almost 30, are of the type "PhanTypeMismatchArgument".
Of these issues, 24 (if I did my math right) are related to the FileBackend class and fixable w/ this patch.

Callers, whether they want an avatar or a gift or award image, already pass in the 2nd param as an int; it only makes sense for them to do so.

Change-Id: I1475b8bb4eb2c9b38bd06387d5c3b1301a1beaf3