Allow passing [ 'raw' => true ] as the parameter to wAvatar#getAvatarURL to get the...avatar URL, without an <img> tag
The getAvatarURL() method is the main public entry point for the wAvatar class.
Despite the name, it acts and has always acted more like getAvatarImgTag() -- the output is *not* just a mere URL for the avatar image, but the whole image wrapped in a HTML-safe <img> tag.
This is great for *most* use cases, but sometimes you do need just the raw URL, e.g. for API modules and the like (I'm thinking of the MediaWikiChat extension here).

Now getAvatarURL() can do exactly what it says on the tin: get and return _just_ the avatar URL and nothing more, nothing less.
This isn't *ideal* but this is by far the least disruptive option.
In practise, this is somewhat of an edge case and most consumers of the class *do* want the <img> wrapping and all; but for the select few that don't, we now have a sensible option.

Change-Id: Ifc60a16507e9b5838579fdd0a8fed9d6f448dd30