Special:EditProfile: Respect profile field visibility settings
The special page Special:EditProfile allows users with the `editothersprofiles`
user right to edit other users' social profiles. However, it currently exposes
the contents of profile fields marked "hidden" or visible only to "friends" or
"friends of friends" when the privileged user isn’t a friend of the user whose
profile they are editing.

Bug: T373265
Change-Id: I4b77ced314bc6cea0ef3657a82e7467d3661fe2a