Remove unused perPage variable from UserBoard.js
The usage of the variable was removed in a3dacb2187076e7439bad9c4b594e47bf0393c25.

The original sajax function wfSendBoardMessage returned the HTML returned by UserBoard#displayMessages.
In that, the $per_page variable was used as the 3rd argument to that call.
The current API module ApiSendUserBoardMessage calls UserBoard#displayMessages but hard-codes in 1 as the 3rd param,
and this has been the case since the introduction of that API module in the aforementioned commit.

Change-Id: I94e022fea8924808a5b0df8cdc763160ccc82a4b