Avoid the $_SESSION superglobal since it won't work as intended in MW 1.45+
This is due to core MW change 59467cab835a75392071aeef93b7f152210d219c/bug T404636.

A bunch of essentially unused, dead code is removed from the uploading special pages (which are largely copies of each other anyway) to remove the $_SESSION use.
Based on some very limited testing on my 1.43 box, this is fine and has zero functional impact, so there's no reason to not do this.

Change-Id: I6c79e977eb24c8a6ddbeff021ca85d675c08f17d